+++ packages/iov-bcp-types/types/atomicswap.d.ts
@@ -10,7 +10,12 @@ export interface SwapData {
     readonly id: SwapIdBytes;
     readonly sender: Address;
     readonly recipient: Address;
-    readonly hashlock: Uint8Array;
+    /**
+     * The hash, whose preimage releases the atomic swap.
+     *
+     * Until we have a way to specify the hashing algirithm, this is SHA256.
+     */
+    readonly hash: Uint8Array;
     readonly amounts: ReadonlyArray<Amount>;
     readonly timeout: number;
     readonly memo?: string;