+++ packages/iov-bcp-types/types/transactions.d.ts
@@ -151,7 +151,8 @@ export interface SwapCounterTransaction extends UnsignedTransaction {
     readonly recipient: Address;
     /** absolute block height at which the counter offer times out */
     readonly timeout: number;
-    readonly hashCode: Uint8Array;
+    /** Copied from the swap offer transaction */
+    readonly hash: Uint8Array;
     readonly memo?: string;
 }
 export interface SwapClaimTransaction extends UnsignedTransaction {